requirements.txt

---

# 📦 requirements.txt

```txt
Django>=5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9

python-dotenv>=1.0

scikit-learn>=1.4
joblib>=1.3
pandas>=2.0
numpy>=1.26

langchain>=0.2
langchain-core>=0.2
langchain-groq>=0.1
